# api/v1/swagger.yaml
...
swagger: "2.0"
...
info:
  version: 1.0.0
  title: Refocus API
  # description: >-
  #   A short description of the application. GFM syntax can be used for rich text
  #   representation.
  # termsOfService: The Terms of Service for the API.
  # contact:
  #   name: The identifying name of the contact person/organization.
  #   url: http://focus.salesforce.com/contact
  #   email: focusapi@salesforce.com
  # license:
  #   name: The license name used for the API.
  #   url: http://focus.salesforce.com/license

# =============================================================================
host: localhost:10010
...
basePath: /v1
...
schemes:
  - https
  - http

# =============================================================================
securityDefinitions:
  jwt:
    type: apiKey
    name: Authorization
    in: header
# =============================================================================
consumes:
  - application/json

# =============================================================================
produces:
  - application/json

# =============================================================================
paths:
...
  /subjects:
    x-swagger-router-controller: subjects
    get:
      security:
        - jwt: []
      summary: Find a subject or subjects
      tags: [ subjects ]
      description: >-
        Find a subject or subjects. You may query using field filters with
        asterisk (*) wildcards. You may also optionally specify sort, limit,
        offset, and a list of fields to include in the response. If the Refocus configuration parameter `useAccessToken` is set to `true`, you must include an `Authorization` request header with your [JSON Web Token](https://tools.ietf.org/html/rfc7519) (JWT) as the value. You can get a token using `POST /v1/register` or `POST /v1/token`.
      externalDocs:
        description: >-
          GET /subjects
        url: http://focus.salesforce.com/docs/api/subjects
      operationId: findSubjects
      parameters:
        -
          $ref: "#/parameters/SubjectsFieldsParam"
        - name: sort
          in: query
          description: >
            Specify the sort order using a field name, e.g. '...?sort=name'. Prepend the
            field name with a minus sign to specify descending order, e.g.
            '...?sort=-name'.
          required: false
          type: array
          collectionFormat: csv
          items:
            type: string
            enum:
              - absolutePath
              - description
              - helpEmail
              - helpUrl
              - id
              - isPublished
              - name
              - parentAbsolutePath
              - parentId
              #- childCount
              - createdBy
              - createdAt
              - updatedAt
              - -absolutePath
              - -description
              - -helpEmail
              - -helpUrl
              - -id
              - -isPublished
              - -name
              - -parentAbsolutePath
              - -parentId
              #- -childCount
              - -createdBy
              - -createdAt
              - -updatedAt
        -
          $ref: "#/parameters/limitParam"
        -
          $ref: "#/parameters/offsetParam"
        -
          name: absolutePath
          in: query
          description: >-
            Filter by absolutePath; asterisk (*) wildcards ok.
          required: false
          type: string
        -
          name: description
          in: query
          description: >-
            Filter by description; asterisk (*) wildcards ok.
          required: false
          type: string
        -
          name: helpEmail
          in: query
          description: >-
            Filter by helpEmail; asterisk (*) wildcards ok.
          required: false
          type: string
        -
          name: helpUrl
          in: query
          description: >-
            Filter by helpUrl; asterisk (*) wildcards ok.
          required: false
          type: string
        -
          name: imageUrl
          in: query
          description: >-
            Filter by imageUrl; asterisk (*) wildcards ok.
          required: false
          type: string
        -
          name: isPublished
          in: query
          description: >-
            Filter by isPublished (true|false).
          required: false
          type: boolean
        -
          name: name
          in: query
          description: >-
            Filter by name; asterisk (*) wildcards ok.
          required: false
          type: string
        -
          name: parentAbsolutePath
          in: query
          description: >-
            Filter by parentAbsolutePath; asterisk (*) wildcards ok.
          required: false
          type: string
      responses:
        200:
          description: >-
            Success, a list of subjects.
          schema:
            type: array
            items:
              $ref: "#/definitions/SubjectsResponse"
        400:
          $ref: "#/responses/400"
        default:
          $ref: "#/responses/genericError"
    post:
      security:
        - jwt: []
      summary: Create a new subject
      tags: [ subjects ]
      description: >-
        Create a new subject. If the Refocus configuration parameter `useAccessToken` is set to `true`, you must include an `Authorization` request header with your [JSON Web Token](https://tools.ietf.org/html/rfc7519) (JWT) as the value. You can get a token using `POST /v1/register` or `POST /v1/token`.
      operationId: postSubject
      parameters:
        -
          name: queryBody
          in: body
          schema:
            type: object
            description: >
              A resource under observation.
            properties:
              description:
                type: string
                maxLength: 4096
                description: >
                  A description of the subject. A lens may choose to display this
                  information to the user.
              helpEmail:
                type: string
                description: >
                  The email address where a user can go to get more help about the
                  subject. A lens may choose to display this information to the user e.g.
                  if the user clicks a “help” icon attached to the subject.
                maxLength: 254
              helpUrl:
                type: string
                maxLength: 2082
                description: >
                  The url where a user can go to get more help about the subject. A lens
                  may choose to display this information to the user e.g. if the user
                  clicks a “help” icon attached to the subject.
              imageUrl:
                type: string
                maxLength: 2082
                description: >
                  Optional icon to represent the subject.
              isPublished:
                type: boolean
                default: false
                description: >
                  Default false. When set to true, the subject is readable by other users;
                  when set to false, the subject is only readable by its owner.
              name:
                type: string
                pattern: ^[0-9A-Za-z_\\-]{1,60}$
                maxLength: 60
                description: >
                  Max 60 characters; case insensitive; allows alpha-numeric characters,
                  underscore (_) and dash (-).
              parentId:
                type: string
                pattern: ^[0-9A-Fa-f]{8}-[0-9A-Fa-f]{4}-[0-9A-Fa-f]{4}-[0-9A-Fa-f]{4}-[0-9A-Fa-f]{12}$
                description: >
                  The absolute path or id of the subject's parent.
              tags:
                type: array
                items:
                  $ref: "#/definitions/TagRequest"
                description: >
                  Tags associated with this model.
              relatedLinks:
                type: array
                items:
                  $ref: "#/definitions/RelatedLinkRequest"
                description: >
                  RelatedLinks associated with this model.
              geolocation:
                type: array
                items:
                  type: number
                description: >
                  If the subject has a physical location, specify its longitude and latitude in this two-element array, e.g. [-122.431297, 37.773972]. The first element in the array represents longitude; the second element represents latitude.
            required:
              - name
      responses:
        201:
          description: >-
            Created.
          schema:
            $ref: "#/definitions/SubjectsResponse"
        400:
          $ref: "#/responses/400"
        403:
          $ref: "#/responses/403"
        default:
          $ref: "#/responses/genericError"

  # ---------------------------------------------------------------------------
  /subjects/{key}:
    x-swagger-router-controller: subjects
    delete:
      security:
        - jwt: []
      summary: Delete the specified subject
      tags: [ subjects ]
      description: >-
        Delete the specified subject. You will not be allowed to delete a
        subject if it has any children subjects under it. Beware! Deleting a
        subject also deletes all the samples associated with this subject. If the Refocus configuration parameter `useAccessToken` is set to `true`, you must include an `Authorization` request header with your [JSON Web Token](https://tools.ietf.org/html/rfc7519) (JWT) as the value. You can get a token using `POST /v1/register` or `POST /v1/token`.
      operationId: deleteSubject
      parameters:
        -
          name: key
          in: path
          description: >-
            The id or absolute path of the subject to delete.
          required: true
          type: string
      responses:
        200:
          description: >-
            Success, returns deleted subject.
          schema:
            $ref: "#/definitions/SubjectsResponse"
        400:
          $ref: "#/responses/400"
        404:
          $ref: "#/responses/404"
        default:
          $ref: "#/responses/genericError"
    get:
      security:
        - jwt: []
      summary: Retrieve the specified subject
      tags: [ subjects ]
      description: >-
        Retrieve the specified subject by its id or name. You may also
        optionally specify a list of fields to include in the response. If the Refocus configuration parameter `useAccessToken` is set to `true`, you must include an `Authorization` request header with your [JSON Web Token](https://tools.ietf.org/html/rfc7519) (JWT) as the value. You can get a token using `POST /v1/register` or `POST /v1/token`.
      operationId: getSubject
      parameters:
        -
          name: key
          in: path
          description: >-
            The id or absolute path of the subject to retrieve.
          required: true
          type: string
        -
          $ref: "#/parameters/SubjectsFieldsParam"
      responses:
        200:
          description: >-
            Success, returns specified subject.
          schema:
            $ref: "#/definitions/SubjectsResponse"
        400:
          $ref: "#/responses/400"
        404:
          $ref: "#/responses/404"
        default:
          $ref: "#/responses/genericError"
    patch:
      security:
        - jwt: []
      summary: Update the specified subject
      tags: [ subjects ]
      description: >-
        Update the specified subject. If a field is not included in the query
        body, that field will not be updated. If the Refocus configuration parameter `useAccessToken` is set to `true`, you must include an `Authorization` request header with your [JSON Web Token](https://tools.ietf.org/html/rfc7519) (JWT) as the value. You can get a token using `POST /v1/register` or `POST /v1/token`.
      operationId: patchSubject
      parameters:
        -
          name: key
          in: path
          description: >-
            The id or absolute path of the subject to update.
          required: true
          type: string
        -
          name: queryBody
          in: body
          required: true
          schema:
            type: object
            description: >
              A resource under observation.
            properties:
              description:
                type: string
                maxLength: 4096
                description: >
                  A description of the subject. A lens may choose to display this
                  information to the user.
              helpEmail:
                type: string
                description: >
                  The email address where a user can go to get more help about the
                  subject. A lens may choose to display this information to the user e.g.
                  if the user clicks a “help” icon attached to the subject.
                maxLength: 254
              helpUrl:
                type: string
                maxLength: 2082
                description: >
                  The url where a user can go to get more help about the subject. A lens
                  may choose to display this information to the user e.g. if the user
                  clicks a “help” icon attached to the subject.
              imageUrl:
                type: string
                maxLength: 2082
                description: >
                  Optional icon to represent the subject.
              isPublished:
                type: boolean
                default: false
                description: >
                  Default false. When set to true, the subject is readable by other users;
                  when set to false, the subject is only readable by its owner.
              name:
                type: string
                pattern: ^[0-9A-Za-z_\\-]{1,60}$
                maxLength: 60
                description: >
                  Max 60 characters; case insensitive; allows alpha-numeric characters,
                  underscore (_) and dash (-).
              parentId:
                type: string
                pattern: ^[0-9A-Fa-f]{8}-[0-9A-Fa-f]{4}-[0-9A-Fa-f]{4}-[0-9A-Fa-f]{4}-[0-9A-Fa-f]{12}$
                description: >
                  The id of the subject's parent.
              tags:
                type: array
                items:
                  $ref: "#/definitions/TagRequest"
                description: >
                  Tags associated with this model.
              relatedLinks:
                type: array
                items:
                  $ref: "#/definitions/RelatedLinkRequest"
                description: >
                  RelatedLinks associated with this model.
              geolocation:
                type: array
                items:
                  type: number
                description: >
                  If the subject has a physical location, specify its longitude and latitude in this two-element array, e.g. [-122.431297, 37.773972]. The first element in the array represents longitude; the second element represents latitude.

      responses:
        200:
          description: >-
            Success, returns updated subject.
          schema:
            $ref: "#/definitions/SubjectsResponse"
        400:
          $ref: "#/responses/400"
        403:
          $ref: "#/responses/403"
        404:
          $ref: "#/responses/404"
        default:
          $ref: "#/responses/genericError"
    put:
      security:
        - jwt: []
      summary: Update the specified subject
      tags: [ subjects ]
      description: >-
        Update the specified subject. If a field is not included in the query
        body, that field will be set to null or its default value. If the Refocus configuration parameter `useAccessToken` is set to `true`, you must include an `Authorization` request header with your [JSON Web Token](https://tools.ietf.org/html/rfc7519) (JWT) as the value. You can get a token using `POST /v1/register` or `POST /v1/token`.
      operationId: putSubject
      parameters:
        -
          name: key
          in: path
          description: >-
            The id or absolute path of the subject to update.
          required: true
          type: string
        -
          name: queryBody
          in: body
          required: true
          schema:
            type: object
            description: >
              A resource under observation.
            properties:
              description:
                type: string
                maxLength: 4096
                description: >
                  A description of the subject. A lens may choose to display this
                  information to the user.
              helpEmail:
                type: string
                description: >
                  The email address where a user can go to get more help about the subject.
                  A lens may choose to display this information to the user e.g. if the
                  user clicks a “help” icon attached to the subject.
                maxLength: 254
              helpUrl:
                type: string
                maxLength: 2082
                description: >
                  The url where a user can go to get more help about the subject. A lens
                  may choose to display this information to the user e.g. if the user
                  clicks a “help” icon attached to the subject.
              imageUrl:
                type: string
                maxLength: 2082
                description: >
                  Optional icon to represent the subject.
              isPublished:
                type: boolean
                default: false
                description: >
                  Default false. When set to true, the subject is readable by other users;
                  when set to false, the subject is only readable by its owner.
              name:
                type: string
                pattern: ^[0-9A-Za-z_\\-]{1,60}$
                maxLength: 60
                description: >
                  Max 60 characters; case insensitive; allows alpha-numeric characters,
                  underscore (_) and dash (-).
              parentId:
                type: string
                pattern: ^[0-9A-Fa-f]{8}-[0-9A-Fa-f]{4}-[0-9A-Fa-f]{4}-[0-9A-Fa-f]{4}-[0-9A-Fa-f]{12}$
                description: >
                  The absolute path or id of the subject's parent.
              tags:
                type: array
                items:
                  $ref: "#/definitions/TagRequest"
                description: >
                  Tags associated with this model.
              relatedLinks:
                type: array
                items:
                  $ref: "#/definitions/RelatedLinkRequest"
                description: >
                  RelatedLinks associated with this model.
              geolocation:
                type: array
                items:
                  type: number
                description: >
                  If the subject has a physical location, specify its longitude and latitude in this two-element array, e.g. [-122.431297, 37.773972]. The first element in the array represents longitude; the second element represents latitude.
            required:
              - name
      responses:
        200:
          description: >-
            Success, returns updated subject.
          schema:
            $ref: "#/definitions/SubjectsResponse"
        400:
          $ref: "#/responses/400"
        403:
          $ref: "#/responses/403"
        404:
          $ref: "#/responses/404"
        default:
          $ref: "#/responses/genericError"

  # ---------------------------------------------------------------------------
  /subjects/{key}/child:
    x-swagger-router-controller: subjects
    post:
      security:
        - jwt: []
      summary: Create a new child subject
      tags: [ subjects ]
      description: >-
        Create a new child subject of the specified subject. If the Refocus configuration parameter `useAccessToken` is set to `true`, you must include an `Authorization` request header with your [JSON Web Token](https://tools.ietf.org/html/rfc7519) (JWT) as the value. You can get a token using `POST /v1/register` or `POST /v1/token`.
      operationId: postChildSubject
      parameters:
        -
          name: key
          in: path
          description: >-
            The id or absolute path of the subject to which you want to add a
            child.
          required: true
          type: string
        -
          name: queryBody
          description: Request body
          in: body
          schema:
            type: object
            description: >
              A resource under observation.
            properties:
              description:
                type: string
                maxLength: 4096
                description: >
                  A description of the subject. A lens may choose to display this
                  information to the user.
              helpEmail:
                type: string
                description: >
                  The email address where a user can go to get more help about the
                  subject. A lens may choose to display this information to the user e.g.
                  if the user clicks a “help” icon attached to the subject.
                maxLength: 254
              helpUrl:
                type: string
                maxLength: 2082
                description: >
                  The url where a user can go to get more help about the subject. A lens
                  may choose to display this information to the user e.g. if the user
                  clicks a “help” icon attached to the subject.
              imageUrl:
                type: string
                maxLength: 2082
                description: >
                  Optional icon to represent the subject.
              isPublished:
                type: boolean
                default: false
                description: >
                  Default false. When set to true, the subject is readable by other users;
                  when set to false, the subject is only readable by its owner.
              name:
                type: string
                pattern: ^[0-9A-Za-z_\\-]{1,60}$
                maxLength: 60
                description: >
                  Max 60 characters; case insensitive; allows alpha-numeric characters,
                  underscore (_) and dash (-).
              tags:
                type: array
                items:
                  $ref: "#/definitions/TagRequest"
                description: >
                  Tags associated with this model.
              relatedLinks:
                type: array
                items:
                  $ref: "#/definitions/RelatedLinkRequest"
                description: >
                  RelatedLinks associated with this model.
            required:
              - name
      responses:
        201:
          description: >-
            Created.
          schema:
            $ref: "#/definitions/SubjectsResponse"
        400:
          $ref: "#/responses/400"
        403:
          $ref: "#/responses/403"
        404:
          $ref: "#/responses/404"
        default:
          $ref: "#/responses/genericError"

  # ---------------------------------------------------------------------------
...
  /subjects/{key}/tags:
    x-swagger-router-controller: subjects
    delete:
      security:
        - jwt: []
      summary: Delete all the tags of the subject
      tags: [ subjects ]
      description: >-
        Delete all the tags associated with the subject. If the Refocus configuration parameter `useAccessToken` is set to `true`, you must include an `Authorization` request header with your [JSON Web Token](https://tools.ietf.org/html/rfc7519) (JWT) as the value. You can get a token using `POST /v1/register` or `POST /v1/token`.
      operationId: deleteSubjectTags
      parameters:
        -
          name: key
          in: path
          description: >-
            The id or absolute path of the subject for which tags needs to be
            deleted.
          required: true
          type: string
      responses:
        200:
          description: >-
            Success, returns updated subject
          schema:
            $ref: "#/definitions/SubjectsResponse"
        400:
          $ref: "#/responses/400"
        404:
          $ref: "#/responses/404"
        default:
          $ref: "#/responses/genericError"

  # ---------------------------------------------------------------------------
  /subjects/{key}/tags/{akey}:
    x-swagger-router-controller: subjects
    delete:
      security:
        - jwt: []
      summary: Delete the specified tag of the subject
      tags: [ subjects ]
      description: >-
        Delete the specified tag of the specified subject. If the Refocus configuration parameter `useAccessToken` is set to `true`, you must include an `Authorization` request header with your [JSON Web Token](https://tools.ietf.org/html/rfc7519) (JWT) as the value. You can get a token using `POST /v1/register` or `POST /v1/token`.
      operationId: deleteSubjectTags
      parameters:
        -
          name: key
          in: path
          description: >-
            The id or absolute path of the subject for which tag needs to be
            deleted.
          required: true
          type: string
        -
          name: akey
          in: path
          description: >-
            The id or absolute path of the tag which needs to be deleted.
          required: true
          type: string
      responses:
        200:
          description: >-
            Success, returns updated subject.
          schema:
            $ref: "#/definitions/SubjectsResponse"
        400:
          $ref: "#/responses/400"
        404:
          $ref: "#/responses/404"
        default:
          $ref: "#/responses/genericError"

  # ---------------------------------------------------------------------------
  /subjects/{key}/relatedLinks:
    x-swagger-router-controller: subjects
    delete:
      security:
        - jwt: []
      summary: Delete all the tags of the subject
      tags: [ subjects ]
      description: >-
        Delete all the tags associated with the subject. If the Refocus configuration parameter `useAccessToken` is set to `true`, you must include an `Authorization` request header with your [JSON Web Token](https://tools.ietf.org/html/rfc7519) (JWT) as the value. You can get a token using `POST /v1/register` or `POST /v1/token`.
      operationId: deleteSubjectRelatedLinks
      parameters:
        -
          name: key
          in: path
          description: >-
            The id or absolute path of the subject for which related links needs
            to be deleted.
          required: true
          type: string
      responses:
        200:
          description: >-
            Success, returns updated subject.
          schema:
            $ref: "#/definitions/SubjectsResponse"
        400:
          $ref: "#/responses/400"
        404:
          $ref: "#/responses/404"
        default:
          $ref: "#/responses/genericError"

  # ---------------------------------------------------------------------------
...
definitions:
...
  TagRequest:
    type: object
    description: >
      Tag that is associated with a model.
    properties:
      name:
        type: string
        maxLength: 60
        description: >
          Name of the tag.
    required:
      - name
...
parameters:
  limitParam:
    name: limit
    in: query
    description: >
        Maximum number of records to return in the response; also useful for
        pagination in the UI.
    required: false
    type: integer
  offsetParam:
    name: offset
    in: query
    description: >
        Offset for paginating the response.
    required: false
    type: integer
...
  SubjectsFieldsParam:
    name: fields
    in: query
    description: >
      Comma-delimited list of field names to include in the response.
    required: false
    type: array
    collectionFormat: csv
    items:
      type: string
      enum:
        - absolutePath
        - description
        - helpEmail
        - helpUrl
        - id
        - imageUrl
        - isPublished
        - name
        - parentAbsolutePath
        - parentId
        #- childCount
        - createdBy
        - createdAt
        - updatedAt
...
responses:
  400:
    description: Bad Request
    schema:
      $ref: '#/definitions/BadRequest'
...
  genericError:
    description: Error
    schema:
      $ref: '#/definitions/ErrorResponse'

# =============================================================================